requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2

# PostgreSQL
sqlalchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
greenlet===3.3.2
psycopg2-binary===2.9.11
# Optional — uncomment when wiring up real push notifications
# firebase-admin==6.5.0
# httpx==0.27.2
